+++ compiler/ml/error_message_utils.ml
@@ -90,6 +90,7 @@ type type_clash_context =
       optional: bool;
     }
   | ArrayValue
+  | TaggedTemplateValue
   | MaybeUnwrapOption
   | IfCondition
   | AssertCondition
@@ -113,30 +114,6 @@ type type_clash_context =
   | ForLoopCondition
   | Await
 
-let context_to_string = function
-  | Some WhileCondition -> "WhileCondition"
-  | Some ForLoopCondition -> "ForLoopCondition"
-  | Some AssertCondition -> "AssertCondition"
-  | Some IfCondition -> "IfCondition"
-  | Some (Statement _) -> "Statement"
-  | Some (MathOperator _) -> "MathOperator"
-  | Some ArrayValue -> "ArrayValue"
-  | Some (SetRecordField _) -> "SetRecordField"
-  | Some (RecordField _) -> "RecordField"
-  | Some MaybeUnwrapOption -> "MaybeUnwrapOption"
-  | Some SwitchReturn -> "SwitchReturn"
-  | Some TryReturn -> "TryReturn"
-  | Some StringConcat -> "StringConcat"
-  | Some (FunctionArgument _) -> "FunctionArgument"
-  | Some JsxComponent -> "JsxComponent"
-  | Some ComparisonOperator -> "ComparisonOperator"
-  | Some IfReturn -> "IfReturn"
-  | Some TernaryReturn -> "TernaryReturn"
-  | Some Await -> "Await"
-  | Some BracedIdent -> "BracedIdent"
-  | Some LetUnwrapReturn -> "LetUnwrapReturn"
-  | None -> "None"
-
 let fprintf = Format.fprintf
 
 let error_type_text ppf type_clash_context =
@@ -145,6 +122,7 @@ let error_type_text ppf type_clash_context =
     | Some (Statement FunctionCall) -> "This function call returns:"
     | Some (MathOperator {is_constant = Some _}) -> "This value has type:"
     | Some ArrayValue -> "This array item has type:"
+    | Some TaggedTemplateValue -> "This interpolated value has type:"
     | Some (SetRecordField _) ->
       "You're assigning something to this field that has type:"
     | Some JsxComponent -> "This JSX tag has type:"
@@ -184,6 +162,8 @@ let error_expected_type_text ppf type_clash_context =
   | Some TernaryReturn -> fprintf ppf "But this ternary is expected to return:"
   | Some ArrayValue ->
     fprintf ppf "But this array is expected to have items of type:"
+  | Some TaggedTemplateValue ->
+    fprintf ppf "But this tag expects interpolations of type:"
   | Some (SetRecordField _) -> fprintf ppf "But the record field is of type:"
   | Some
       (RecordField {field_name = "children"; jsx = Some {jsx_type = `Fragment}})