+++ packages/@rescript/runtime/lib/es6/Stdlib_TaggedTemplate.mjs
@@ -1,0 +1,9 @@
+
+
+
+let make = (transform => (strings, ...values) => transform(strings, values));
+
+export {
+  make,
+}
+/* No side effect */